+++ src/sc/index.d.ts
@@ -13,6 +13,7 @@ export class ContractParam {
 
   static string(value: string): ContractParam
   static boolean(value: any): ContractParam
+  static hash160(value: string): ContractParam
   static integer(value: any): ContractParam
   static byteArray(value: any, format: string): ContractParam
   static array(param: any[]): ContractParam