+++ src/utils.js
@@ -29,9 +29,7 @@ export const str2ab = str => {
  * @returns {number[]}
  */
 export const hexstring2ab = str => {
-  if (typeof str !== 'string') {
-    throw new Error('hexstring2ab expects a string')
-  }
+  ensureHex(str)
   if (!str.length) return new Uint8Array()
   const iters = str.length / 2
   const result = new Uint8Array(iters)
@@ -123,8 +121,7 @@ export const num2fixed8 = (num, size = 8) => {
  * @return {number}
  */
 export const fixed82num = (fixed8hex) => {
-  if (typeof fixed8hex !== 'string') throw new Error('fixed8hex must be a string')
-  if (fixed8hex.length % 2 !== 0) throw new Error('fixed8hex must be hex')
+  ensureHex(fixed8hex)
   if (fixed8hex === '') return 0
   return Fixed8.fromReverseHex(fixed8hex).toNumber()
 }
@@ -156,9 +153,9 @@ export const num2VarInt = (num) => {
  * @returns {string} XOR output as a HEX string
  */
 export const hexXor = (str1, str2) => {
-  if (typeof str1 !== 'string' || typeof str2 !== 'string') throw new Error('hexXor expects hex strings')
+  ensureHex(str1)
+  ensureHex(str2)
   if (str1.length !== str2.length) throw new Error('strings are disparate lengths')
-  if (str1.length % 2 !== 0) throw new Error('strings must be hex')
   const result = []
   for (let i = 0; i < str1.length; i += 2) {
     result.push(parseInt(str1.substr(i, 2), 16) ^ parseInt(str2.substr(i, 2), 16))
@@ -189,13 +186,37 @@ export const reverseArray = arr => {
  * @return {string} HEX string reversed in 2s.
  */
 export const reverseHex = hex => {
-  if (typeof hex !== 'string') throw new Error('reverseHex expects a string')
-  if (hex.length % 2 !== 0) throw new Error(`Incorrect Length: ${hex}`)
+  ensureHex(hex)
   let out = ''
   for (let i = hex.length - 2; i >= 0; i -= 2) {
     out += hex.substr(i, 2)
   }
   return out
+}
+
+const hexRegex = /^([0-9A-Fa-f]{2})*$/
+
+/**
+ * Checks if input is a hexstring. Empty string is considered a hexstring.
+ * @example
+ * isHex('0101') = true
+ * isHex('') = true
+ * isHex('0x01') = false
+ * @param {string} str
+ * @return {boolean}
+ */
+export const isHex = str => {
+  try {
+    return hexRegex.test(str)
+  } catch (err) { return false }
+}
+
+/**
+ * Throws an error if input is not hexstring.
+ * @param {string} str
+ */
+export const ensureHex = str => {
+  if (!isHex(str)) throw new Error(`Expected a hexstring but got ${str}`)
 }
 
 /**